refactor: move CMS MetaTags to the common folder
